refactoring day1 files
