+++ src/main/resources/inputDay1.txt
